separate components code
