+++ src/components/burger-ingredients/tab-bar/action-tab/action-tab.module.css
@@ -1,0 +1,3 @@
+.tabLink {
+  composes: anchor from global;
+}
