+++ src/components/burger-ingredients/tab-bar/tab-bar.jsx
@@ -1,31 +1,11 @@
 // libraries
 import React from "react";
-import PropTypes from "prop-types";
 
 // components
-import { Tab } from "@ya.praktikum/react-developer-burger-ui-components"
+import ActionTab from "./action-tab/action-tab";
 
 // styles
 import styles from "./tab-bar.module.css";
-
-
-
-function ActionTab({ value, title, onClick, currentlyActive }) {
-  return (
-    <Tab value={value} active={currentlyActive === value} onClick={onClick}>
-      <a href={`#${value}`} className={styles.tabLink}>{title}</a>
-    </Tab>  
-  );
-};
-
-ActionTab.propTypes = PropTypes.exact(
-  {
-    value: PropTypes.string.isRequired,
-    title: PropTypes.string.isRequired,
-    onClick: PropTypes.func.isRequired,
-    currentlyActive: PropTypes.string.isRequired
-  }
-).isRequired;
