+++ src/components/burger-ingredients/tab-bar/tab-bar.module.css
@@ -1,7 +1,3 @@
-.tabLink {
-  composes: anchor from global;
-}
-
 .tabBar {
   composes: mb-10 from global;
   margin-top: 0;
